feat: add `<SpeakerSocials>`
