+++ src/app/components/Speakers/SpeakerCard.tsx
@@ -1,13 +1,6 @@
 import { TSpeakerCard } from "@/types/home/speakerCard.type";
 import Image from "next/image";
-import Link from "next/link";
-import {
-  RiGithubLine,
-  RiGlobalLine,
-  RiInstagramLine,
-  RiLinkedinLine,
-  RiTwitterXLine,
-} from "react-icons/ri";
+import SpeakerSocials from "./SpeakerSocials";
 
 const SpeakerCard = (props: TSpeakerCard) => {
   return (
@@ -30,25 +23,5 @@ const SpeakerCard = (props: TSpeakerCard) => {
     </div>
   );
 };
-
-const SpeakerSocials = (socials: TSpeakerCard["socials"]) => {
-  return Object.entries(socials).map(([icon, link]) => {
-    const Icon = socialIcons[icon as keyof typeof socialIcons];
-
-    return (
-      <Link key={icon} href={link} target="_blank">
-        <Icon className="size-5 fill-zinc-400 transition-all duration-200 hover:-translate-y-0.5 hover:fill-orange-500" />
-      </Link>
-    );
-  });
-};
-
-export const socialIcons = {
-  github: RiGithubLine,
-  twitter: RiTwitterXLine,
-  linkedin: RiLinkedinLine,
-  instagram: RiInstagramLine,
-  website: RiGlobalLine,
-} as const;
 
 export default SpeakerCard;
